feat(vertical-menu-full-screen): add slide in animation on menu open/close
