+++ src/components/vertical-menu/vertical-menu-full-screen/vertical-menu-full-screen.test.tsx
@@ -1,10 +1,13 @@
 import React from "react";
-import { render, screen } from "@testing-library/react";
+import { render, screen, act } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
 import Portal from "../../portal";
 import FocusTrap from "../../../__internal__/focus-trap/focus-trap.component";
 import { VerticalMenuFullScreen, VerticalMenuItem } from "..";
+import useMediaQuery from "../../../hooks/useMediaQuery";
+
+jest.mock("../../../hooks/useMediaQuery");
 
 jest.mock("../../portal", () =>
   jest.fn(({ children }) => <div>{children}</div>),
@@ -61,47 +64,24 @@ describe("VerticalMenuFullScreen", () => {
     expect(FocusTrap).toHaveBeenCalled();
   });
 
-  // TODO remove skip as part of FE-5650
-  // eslint-disable-next-line jest/no-disabled-tests
-  it.skip('should render hidden when "isOpen" prop is false', () => {
+  test('should not render the menu when "isOpen" prop is false', () => {
     render(
       <VerticalMenuFullScreen isOpen={false} onClose={() => {}}>
         <VerticalMenuItem title="Item1" />
       </VerticalMenuFullScreen>,
     );
 
-    expect(screen.getByRole("navigation", { hidden: true })).toHaveStyle({
-      visibility: "hidden",
-      transform: "translateX(-100%)",
-    });
+    expect(screen.queryByRole("navigation")).not.toBeInTheDocument();
   });
 
-  // TODO remove skip as part of FE-5650
-  // eslint-disable-next-line jest/no-disabled-tests
-  it.skip('should render visible when "isOpen" prop is true', () => {
+  test('should render visible when "isOpen" prop is true', () => {
     render(
       <VerticalMenuFullScreen isOpen onClose={() => {}}>
         <VerticalMenuItem title="Item1" />
       </VerticalMenuFullScreen>,
     );
 
-    expect(screen.getByRole("navigation")).toHaveStyle({
-      visibility: "visible",
-      transform: "translateX(0)",
-    });
-  });
-
-  // TODO remove this as part of FE-5650
-  it("should not render the menu if isOpen is false", () => {
-    render(
-      <VerticalMenuFullScreen isOpen={false} onClose={() => {}}>
-        <VerticalMenuItem title="Item1" />
-      </VerticalMenuFullScreen>,
-    );
-
-    const menu = screen.queryByRole("navigation");
-
-    expect(menu).not.toBeInTheDocument();
+    expect(screen.getByRole("navigation")).toBeVisible();
   });
 
   it("should override the scrollbar styling", () => {
@@ -160,6 +140,42 @@ describe("VerticalMenuFullScreen", () => {
     expect(onClose).toHaveBeenCalled();
   });
 
+  it("should invoke onClose callback when Space is pressed on the close Icon", async () => {
+    const onClose = jest.fn();
+    const user = userEvent.setup();
+    render(
+      <VerticalMenuFullScreen isOpen onClose={onClose}>
+        <VerticalMenuItem title="Item1" />
+      </VerticalMenuFullScreen>,
+    );
+
+    const closeIcon = screen.getByRole("button", { name: "Close" });
+    act(() => {
+      closeIcon.focus();
+    });
+    await user.keyboard(" ");
+
+    expect(onClose).toHaveBeenCalled();
+  });
+
+  it("should invoke onClose callback when Enter is pressed on the close Icon", async () => {
+    const onClose = jest.fn();
+    const user = userEvent.setup();
+    render(
+      <VerticalMenuFullScreen isOpen onClose={onClose}>
+        <VerticalMenuItem title="Item1" />
+      </VerticalMenuFullScreen>,
+    );
+
+    const closeIcon = screen.getByRole("button", { name: "Close" });
+    act(() => {
+      closeIcon.focus();
+    });
+    await user.keyboard("{Enter}");
+
+    expect(onClose).toHaveBeenCalled();
+  });
+
   it("should have the expected data attributes", () => {
     render(
       <VerticalMenuFullScreen
@@ -177,5 +193,39 @@ describe("VerticalMenuFullScreen", () => {
     expect(menu).toHaveAttribute("data-component", "vertical-menu-full-screen");
     expect(menu).toHaveAttribute("data-element", "foo");
     expect(menu).toHaveAttribute("data-role", "bar");
+  });
+
+  it("should not render with animation when reduce motion is set", () => {
+    const mockedUseMediaQuery = jest.mocked(useMediaQuery);
+    mockedUseMediaQuery.mockReturnValue(false);
+
+    render(
+      <VerticalMenuFullScreen isOpen onClose={() => {}}>
+        <VerticalMenuItem title="Item1" />
+      </VerticalMenuFullScreen>,
+    );
+
+    const menu = screen.getByRole("navigation");
+
+    expect(menu).not.toHaveStyleRule("transition", "all 0.3s ease");
+
+    mockedUseMediaQuery.mockReset();
+  });
+
+  it("should render with animation when reduce motion is not set", () => {
+    const mockedUseMediaQuery = jest.mocked(useMediaQuery);
+    mockedUseMediaQuery.mockReturnValue(true);
+
+    render(
+      <VerticalMenuFullScreen isOpen onClose={() => {}}>
+        <VerticalMenuItem title="Item1" />
+      </VerticalMenuFullScreen>,
+    );
+
+    const menu = screen.getByRole("navigation");
+
+    expect(menu).toHaveStyleRule("transition", "all 0.3s ease");
+
+    mockedUseMediaQuery.mockReset();
   });
 });
