+++ src/components/vertical-menu/vertical-menu.style.ts
@@ -114,7 +114,12 @@ export const StyledVerticalMenu = styled(StyledBox)`
   }
 `;
 
-export const StyledVerticalMenuFullScreen = styled(StyledBox)`
+interface FullScreenProps {
+  isOpen: boolean;
+  prefersReducedMotion?: boolean;
+}
+
+export const StyledVerticalMenuFullScreen = styled(StyledBox)<FullScreenProps>`
   position: fixed;
   top: 0;
   bottom: 0;
@@ -125,8 +130,27 @@ export const StyledVerticalMenuFullScreen = styled(StyledBox)`
   overflow: auto;
   background-color: var(--colorsComponentsLeftnavWinterStandardBackground);
   box-sizing: border-box;
-  transition: all 0.3s ease;
   z-index: ${({ theme }) => theme.zIndex.fullScreenModal};
+
+  ${({ prefersReducedMotion }) =>
+    !prefersReducedMotion &&
+    css`
+      transition: all 0.3s ease;
+    `}
+
+  ${({ isOpen }) =>
+    isOpen &&
+    css`
+      visibility: visible;
+      transform: translateX(0);
+    `}
+
+  ${({ isOpen }) =>
+    !isOpen &&
+    css`
+      transform: translateX(-100%);
+      visibility: hidden;
+    `}
 
   // TODO remove hardcoded values when DS have had chance to review which token to use
   &::-webkit-scrollbar-track {
