+++ src/components/vertical-menu/vertical-menu.test.tsx
@@ -1,13 +1,7 @@
 import React from "react";
-import { render, screen, act } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
+import { render, screen } from "@testing-library/react";
 
-import {
-  VerticalMenu,
-  VerticalMenuFullScreen,
-  VerticalMenuItem,
-  VerticalMenuTrigger,
-} from ".";
+import { VerticalMenu, VerticalMenuItem } from ".";
 
 describe("VerticalMenu", () => {
   it("should accept aria-label prop", () => {
@@ -100,108 +94,5 @@ describe("VerticalMenu", () => {
     expect(menu).toHaveAttribute("data-component", "vertical-menu");
     expect(menu).toHaveAttribute("data-element", "foo");
     expect(menu).toHaveAttribute("data-role", "bar");
-  });
-
-  describe("callbacks", () => {
-    it(`should close the Vertical Menu Full Screen when escape key is pressed`, async () => {
-      const user = userEvent.setup();
-
-      const setMockState = jest.fn();
-      const useStateMock: any = (useState: any) => [useState, setMockState];
-      jest.spyOn(React, "useState").mockImplementation(useStateMock);
-
-      render(
-        <>
-          <VerticalMenuFullScreen isOpen onClose={setMockState(false)}>
-            <VerticalMenuItem title="Item1" />
-          </VerticalMenuFullScreen>
-          ,
-        </>,
-      );
-      await user.keyboard("{esc}");
-      expect(setMockState).toHaveBeenCalledWith(false);
-    });
-
-    it(`should call onClick callback when a click event is triggered and using VerticalMenuTrigger`, async () => {
-      const user = userEvent.setup();
-
-      const setMockState = jest.fn();
-      const useStateMock: any = (useState: any) => [useState, setMockState];
-      jest.spyOn(React, "useState").mockImplementation(useStateMock);
-
-      render(
-        <VerticalMenuTrigger onClick={setMockState(false)}>
-          Menu
-        </VerticalMenuTrigger>,
-      );
-      await user.click(screen.getByText("Menu"));
-      expect(setMockState).toHaveBeenCalledWith(false);
-    });
-
-    it(`should call onClose callback when a click event is triggered`, async () => {
-      const user = userEvent.setup();
-
-      const setMockState = jest.fn();
-      const useStateMock: any = (useState: any) => [useState, setMockState];
-      jest.spyOn(React, "useState").mockImplementation(useStateMock);
-
-      render(
-        <>
-          <VerticalMenuFullScreen isOpen onClose={setMockState(false)}>
-            <VerticalMenuItem title="Item1" />
-          </VerticalMenuFullScreen>
-          ,
-        </>,
-      );
-      const closeBtn = screen.getByRole("button");
-      await user.click(closeBtn);
-      expect(setMockState).toHaveBeenCalledWith(false);
-    });
-
-    it(`should call onClose callback when a Space key event is triggered`, async () => {
-      const user = userEvent.setup();
-
-      const setMockState = jest.fn();
-      const useStateMock: any = (useState: any) => [useState, setMockState];
-      jest.spyOn(React, "useState").mockImplementation(useStateMock);
-
-      render(
-        <>
-          <VerticalMenuFullScreen isOpen onClose={setMockState(false)}>
-            <VerticalMenuItem title="Item1" />
-          </VerticalMenuFullScreen>
-          ,
-        </>,
-      );
-      const closeBtn = screen.getByRole("button");
-      act(() => {
-        closeBtn.focus();
-      });
-      await user.keyboard("{Space}");
-      expect(setMockState).toHaveBeenCalledWith(false);
-    });
-
-    it(`should call onClose callback when a Enter key event is triggered`, async () => {
-      const user = userEvent.setup();
-
-      const setMockState = jest.fn();
-      const useStateMock: any = (useState: any) => [useState, setMockState];
-      jest.spyOn(React, "useState").mockImplementation(useStateMock);
-
-      render(
-        <>
-          <VerticalMenuFullScreen isOpen onClose={setMockState(false)}>
-            <VerticalMenuItem title="Item1" />
-          </VerticalMenuFullScreen>
-          ,
-        </>,
-      );
-      const closeBtn = screen.getByRole("button");
-      act(() => {
-        closeBtn.focus();
-      });
-      await user.keyboard("{enter}");
-      expect(setMockState).toHaveBeenCalledWith(false);
-    });
   });
 });
